python-data-engineer/week12/exercises/query_acceleration.py
"""
Query acceleration exercises.
"""

from typing import Dict, Any, Optional, List, Union
import logging
from datetime import datetime, timedelta
from pyspark.sql import SparkSession
from pyspark.sql.functions import *

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class MaterializedView:
    """Materialized view management."""
    
    def __init__(self, spark: SparkSession):
        """Initialize view manager."""
        self.spark = spark
        self.views = {}
        self.refresh_intervals = {}
    
    def create_view(
        self,
        name: str,
        query: str,
        refresh_interval: timedelta = timedelta(hours=1)
    ) -> None:
        """
        Create materialized view.
        
        Args:
            name: View name
            query: SQL query
            refresh_interval: Refresh interval
        """
        try:
            # Create view
            df = self.spark.sql(query)
            df.createOrReplaceTempView(name)
            
            # Store metadata
            self.views[name] = {
                'query': query,
                'last_refresh': datetime.now(),
                'columns': df.columns
            }
            
            self.refresh_intervals[name] = refresh_interval
            
            logger.info(f"Created view: {name}")
            
        except Exception as e:
            logger.error(f"Failed to create view: {e}")
            raise
    
    def refresh_view(
        self,
        name: str,
        force: bool = False
    ) -> None:
        """
        Refresh materialized view.
        
        Args:
            name: View name
            force: Force refresh
        """
        try:
            if name not in self.views:
                raise ValueError(f"View not found: {name}")
            
            # Check if refresh is needed
            last_refresh = self.views[name]['last_refresh']
            interval = self.refresh_intervals[name]
            
            if force or datetime.now() - last_refresh > interval:
                # Refresh view
                df = self.spark.sql(self.views[name]['query'])
                df.createOrReplaceTempView(name)
                
                # Update metadata
                self.views[name]['last_refresh'] = datetime.now()
                
                logger.info(f"Refreshed view: {name}")
            
        except Exception as e:
            logger.error(f"Failed to refresh view: {e}")
            raise
    
    def get_view_stats(
        self,
        name: str
    ) -> Dict[str, Any]:
        """
        Get view statistics.
        
        Args:
            name: View name
        
        Returns:
            View statistics
        """
        try:
            if name not in self.views:
                raise ValueError(f"View not found: {name}")
            
            return self.views[name]
            
        except Exception as e:
            logger.error(f"Failed to get view stats: {e}")
            raise

class AggregationTable:
    """Aggregation table management."""
    
    def __init__(self, spark: SparkSession):
        """Initialize table manager."""
        self.spark = spark
        self.tables = {}
    
    def create_table(
        self,
        name: str,
        source_table: str,
        dimensions: List[str],
        measures: List[str],
        where_clause: Optional[str] = None
    ) -> None:
        """
        Create aggregation table.
        
        Args:
            name: Table name
            source_table: Source table
            dimensions: Dimension columns
            measures: Measure columns
            where_clause: WHERE clause
        """
        try:
            # Build query
            select_clause = ", ".join(
                dimensions + [f"SUM({m}) as {m}" for m in measures]
            )
            
            query = f"""
                SELECT {select_clause}
                FROM {source_table}
                {f'WHERE {where_clause}' if where_clause else ''}
                GROUP BY {', '.join(dimensions)}
            """
            
            # Create table
            self.spark.sql(query).createOrReplaceTempView(name)
            
            # Store metadata
            self.tables[name] = {
                'source': source_table,
                'dimensions': dimensions,
                'measures': measures,
                'where': where_clause,
                'created_at': datetime.now()
            }
            
            logger.info(f"Created table: {name}")
            
        except Exception as e:
            logger.error(f"Failed to create table: {e}")
            raise
    
    def refresh_table(
        self,
        name: str
    ) -> None:
        """
        Refresh aggregation table.
        
        Args:
            name: Table name
        """
        try:
            if name not in self.tables:
                raise ValueError(f"Table not found: {name}")
            
            metadata = self.tables[name]
            
            # Recreate table
            self.create_table(
                name,
                metadata['source'],
                metadata['dimensions'],
                metadata['measures'],
                metadata['where']
            )
            
        except Exception as e:
            logger.error(f"Failed to refresh table: {e}")
            raise

class DynamicFilter:
    """Dynamic filter management."""
    
    def __init__(self, spark: SparkSession):
        """Initialize filter manager."""
        self.spark = spark
        self.filters = {}
    
    def create_filter(
        self,
        name: str,
        table: str,
        column: str,
        condition: str,
        update_freq: timedelta = timedelta(minutes=5)
    ) -> None:
        """
        Create dynamic filter.
        
        Args:
            name: Filter name
            table: Table name
            column: Column name
            condition: Filter condition
            update_freq: Update frequency
        """
        try:
            # Create filter
            query = f"""
                SELECT DISTINCT {column}
                FROM {table}
                WHERE {condition}
            """
            
            df = self.spark.sql(query)
            df.createOrReplaceTempView(f"{name}_filter")
            
            # Store metadata
            self.filters[name] = {
                'table': table,
                'column': column,
                'condition': condition,
                'update_freq': update_freq,
                'last_update': datetime.now()
            }
            
            logger.info(f"Created filter: {name}")
            
        except Exception as e:
            logger.error(f"Failed to create filter: {e}")
            raise
    
    def apply_filter(
        self,
        name: str,
        query: str
    ) -> str:
        """
        Apply dynamic filter.
        
        Args:
            name: Filter name
            query: SQL query
        
        Returns:
            Modified query
        """
        try:
            if name not in self.filters:
                raise ValueError(f"Filter not found: {name}")
            
            # Check if update is needed
            metadata = self.filters[name]
            if datetime.now() - metadata['last_update'] > metadata['update_freq']:
                self.update_filter(name)
            
            # Modify query
            filter_condition = f"""
                {metadata['column']} IN (
                    SELECT {metadata['column']}
                    FROM {name}_filter
                )
            """
            
            if 'WHERE' in query.upper():
                modified = query.replace(
                    'WHERE',
                    f'WHERE {filter_condition} AND'
                )
            else:
                modified = f"{query} WHERE {filter_condition}"
            
            return modified
            
        except Exception as e:
            logger.error(f"Failed to apply filter: {e}")
            raise
    
    def update_filter(
        self,
        name: str
    ) -> None:
        """
        Update dynamic filter.
        
        Args:
            name: Filter name
        """
        try:
            metadata = self.filters[name]
            
            # Recreate filter
            self.create_filter(
                name,
                metadata['table'],
                metadata['column'],
                metadata['condition'],
                metadata['update_freq']
            )
            
        except Exception as e:
            logger.error(f"Failed to update filter: {e}")
            raise 